Move Tester to lib.rs and add some asserts to tests
